Organize model files into C, L, S folders
