+++ lib/layout/dummyLayout.js
@@ -8,7 +8,7 @@
  * @author John O'Brien.
  */
 
-var CardinalPortPosition = require('./../model/cardinalPortPosition');
+var CardinalPortPosition = require('./../model/layout/cardinalPortPosition');
 var CardinalDirection = require('./../enum/cardinalDirection');
