+++ lib/layout/forceDirectedLayout.js
@@ -7,7 +7,7 @@
  * @author John O'Brien.
  */
 
-var CardinalPortPosition = require('./../model/cardinalPortPosition');
+var CardinalPortPosition = require('./../model/layout/cardinalPortPosition');
 
 var change = false;
 //ideal length
