+++ lib/model/content/edge.js
@@ -1,8 +1,8 @@
-var Lang = require('./../utilities/lang');
-var AttrValidators = require('./../utilities/attrValidators');
-var createClass = require('./../utilities/createClass');
+var Lang = require('./../../utilities/lang');
+var AttrValidators = require('./../../utilities/attrValidators');
+var createClass = require('./../../utilities/createClass');
 var GraphModelBase = require('./graphModelBase');
-var EdgeLayout = require('./edgeLayout');
+var EdgeLayout = require('./../layout/edgeLayout');
 
 var Port = require('./port');
