+++ lib/model/content/gclEditorModel.js
@@ -1,10 +1,10 @@
-var createClass = require('./../utilities/createClass');
-var AttrValidators = require('./../utilities/attrValidators');
-var Lang = require('./../utilities/lang');
+var createClass = require('./../../utilities/createClass');
+var AttrValidators = require('./../../utilities/attrValidators');
+var Lang = require('./../../utilities/lang');
 var Graph = require('./graph');
-var parse = require('./../parse/parse');
-var ParseError = require('./../parse/parseError');
-var autoLayout = require('./../layout/forceDirectedLayout');
+var parse = require('./../../parse/parse');
+var ParseError = require('./../../parse/parseError');
+var autoLayout = require('./../../layout/forceDirectedLayout');
 
 var GCLEditorModel = createClass({
