+++ lib/model/content/graph.js
@@ -1,6 +1,6 @@
-var createClass = require('./../utilities/createClass');
+var createClass = require('./../../utilities/createClass');
 var GraphModelBase = require('./graphModelBase');
-var AttrValidators = require('./../utilities/attrValidators');
+var AttrValidators = require('./../../utilities/attrValidators');
 
 var Vertex = require('./vertex');
 var Edge = require('./edge');
