+++ lib/model/content/graphModelBase.js
@@ -1,6 +1,6 @@
-var Lang = require('../utilities/lang');
-var ObjectUtils = require('../utilities/objects');
-var createClass = require('./../utilities/createClass');
+var Lang = require('../../utilities/lang');
+var ObjectUtils = require('../../utilities/objects');
+var createClass = require('../../utilities/createClass');
 
 var SCOPE_SYMBOLS = {
   NODE: '>',
