+++ lib/model/content/port.js
@@ -1,12 +1,12 @@
-var Lang = require('./../utilities/lang');
-var AttrValidators = require('./../utilities/attrValidators');
-var createClass = require('./../utilities/createClass');
+var Lang = require('./../../utilities/lang');
+var AttrValidators = require('./../../utilities/attrValidators');
+var createClass = require('./../../utilities/createClass');
 var GraphModelBase = require('./graphModelBase');
-var PortType = require('./../enum/portType');
-var PositionUtils = require('../utilities/positionUtils.js');
+var PortType = require('./../../enum/portType');
+var PositionUtils = require('../../utilities/positionUtils.js');
 
-var CardinalPortPosition = require('./cardinalPortPosition');
-var CardinalDirection = require('./../enum/cardinalDirection');
+var CardinalPortPosition = require('../layout/cardinalPortPosition');
+var CardinalDirection = require('../../enum/cardinalDirection');
 
 /**
  * Definition for the Port class. Ports are the points on a Vertex where
