+++ lib/model/content/vertex.js
@@ -1,14 +1,14 @@
-var Lang = require('./../utilities/lang');
-var createClass = require('./../utilities/createClass');
-var AttrValidators = require('./../utilities/attrValidators');
+var Lang = require('./../../utilities/lang');
+var createClass = require('./../../utilities/createClass');
+var AttrValidators = require('./../../utilities/attrValidators');
 var GraphModelBase = require('./graphModelBase');
-var PortType = require('./../enum/portType');
-var VertexStyles = require('../styles/vertexStyles');
+var PortType = require('./../../enum/portType');
+var VertexStyles = require('../style/vertexStyles');
 
 var Port = require('./port');
 
-var CardinalPortPosition = require('./cardinalPortPosition');
-var CardinalDirection = require('./../enum/cardinalDirection');
+var CardinalPortPosition = require('../layout/cardinalPortPosition');
+var CardinalDirection = require('../../enum/cardinalDirection');
 
 var validateArrayOfPortsOfType = function(portType) {
   return function(val) {
