+++ lib/model/layout/cardinalPortPosition.js
@@ -1,7 +1,7 @@
-var Lang = require('./../utilities/lang');
-var AttrValidators = require('./../utilities/attrValidators');
-var createClass = require('./../utilities/createClass');
-var CardinalDirection = require('../enum/cardinalDirection');
+var Lang = require('../../utilities/lang');
+var AttrValidators = require('../../utilities/attrValidators');
+var createClass = require('../../utilities/createClass');
+var CardinalDirection = require('../../enum/cardinalDirection');
 
 var DIRECTION_TO_ANGLE = {};
 DIRECTION_TO_ANGLE[CardinalDirection.NORTH] = Math.PI / 2;
