+++ lib/model/layout/edgeLayout.js
@@ -1,6 +1,6 @@
-var createClass = require('../utilities/createClass');
+var createClass = require('../../utilities/createClass');
 var ReroutePointsList = require('./reroutePointsList');
-var AttrValidators = require('./../utilities/attrValidators');
+var AttrValidators = require('../../utilities/attrValidators');
 
 var EdgeLayout = createClass({
   instance: {
