+++ lib/model/layout/reroutePoint.js
@@ -1,6 +1,6 @@
-var createClass = require('../utilities/createClass');
-var Lang = require('../utilities/lang');
-var translateCoordinates = require('../utilities/translateCoordinates');
+var createClass = require('../../utilities/createClass');
+var Lang = require('../../utilities/lang');
+var translateCoordinates = require('../../utilities/translateCoordinates');
 
 var ORIGIN = {x: 0, y: 0};
