+++ lib/model/layout/reroutePointsList.js
@@ -1,5 +1,5 @@
-var createClass = require('../utilities/createClass');
-var ArrayList = require('./arrayList');
+var createClass = require('../../utilities/createClass');
+var ArrayList = require('../arrayList');
 
 var ReroutePointsList = createClass({
   extend: ArrayList,
