+++ lib/model/style/styles.js
@@ -1,4 +1,4 @@
-var createClass = require('./../utilities/createClass');
+var createClass = require('../../utilities/createClass');
 
 var Styles = createClass({
