+++ lib/model/style/vertexStyles.js
@@ -1,5 +1,5 @@
-var Lang = require('./../utilities/lang');
-var createClass = require('./../utilities/createClass');
+var Lang = require('../../utilities/lang');
+var createClass = require('../../utilities/createClass');
 var Styles = require('./styles');
 
 var VertexStyles = createClass({
