+++ lib/parse/modelFactory.js
@@ -1,7 +1,7 @@
-var Graph = require('../model/graph');
-var Edge = require('../model/edge');
-var Vertex = require('../model/vertex');
-var Port = require('../model/port');
+var Graph = require('../model/content/graph');
+var Edge = require('../model/content/edge');
+var Vertex = require('../model/content/vertex');
+var Port = require('../model/content/port');
 var PortType = require('../enum/portType');
 
 var ObjectUtils = require('../utilities/objects');
