+++ lib/utilities/positionUtils.js
@@ -1,5 +1,5 @@
 var CardinalDirection = require('../enum/cardinalDirection');
-var CardinalPortPosition = require('../model/cardinalPortPosition');
+var CardinalPortPosition = require('../model/layout/cardinalPortPosition');
 /**
  * @class PositionUtils
  * @static
